third_party/WebKit/LayoutTests/platform/win/bluetooth/server/disconnect/disconnect-once-expected.txt

This is a testharness.js-based test.
PASS 'connected' is set to false after disconnect is called. 
Harness: the test ran to completion.
